+++ CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `min_process_steps_per_inference` in `ProcessConfigModel` is now `min_frac_process_responses_per_collection` and dynamically determines the minimum process responses (what were previously called process steps) to return based on the specified fraction of however many processes are currently open. The value is validated to be between 0 and 1 (inclusive).
 - The terminal keypresses 'j' and 'l' now update `min_frac_process_responses_per_collection` to use one process more/fewer than what the previous value did.
 - `instance_launch_delay` has been renamed to `launch_delay` for simplicity.
+- Switched to mio (using `UdpSocket`s) in Rust instead of using `selectors` and `socket` in Python for multiplexing and synchronization between EPI and env processes.
 
 ### Removed